pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "assistant"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
  "chainlit>=0.7.600",
  "langchain>=0.1.16",
  "langchain-community>=0.0.29",
  "pydantic>=2.6",
  "pdf2image",
  "pytesseract",
  "Pillow",
  "python-dotenv",
  "chromadb",
]

[tool.setuptools.packages.find]
include = ["assistant"]
exclude = ["db*", "data*", "images*", "preview*"]